+++ hdl/formal/catalogue/sc_pernarowski.v
@@ -159,11 +159,11 @@ always @(posedge clk or negedge rst_n) begin
         if (((v_next >= P_V_THRESHOLD)) && !_thr_prev) begin
             spike_out <= 1'b1;
             v_reg <= v_next;
+            v_out <= v_next;
             w_reg <= w_next;
+            w_out <= w_next;
             z_reg <= z_next;
-            v_out <= v_reg;
-            w_out <= w_reg;
-            z_out <= z_reg;
+            z_out <= z_next;
         end else begin
             spike_out <= 1'b0;
             v_reg <= v_next;